bug with unsigned int pgsql in Root 5.34.36 - unsigned int is bigint in db now

+++ uni_db/scheme/uni_db.sql
@@ -21,7 +21,7 @@
 -- replace true : 1
 -- replace bytea : blob
 
-CREATE DOMAIN uint AS int;
+CREATE DOMAIN uint AS bigint;
 
 -- EXPERIMENT RUN PERIODS (SESSIONS)
 create table run_period
